refactor: use internal product service directly
