+++ apps/product-query-svc/adapters/inmem/product_repository.go
@@ -5,27 +5,27 @@ import (
 	"strings"
 	"sync"
 
-	appsdomain "github.com/fightingBald/GoTuto/apps/product-query-svc/domain"
-	appsports "github.com/fightingBald/GoTuto/apps/product-query-svc/ports"
+	"github.com/fightingBald/GoTuto/internal/domain"
+	"github.com/fightingBald/GoTuto/internal/ports"
 )
 
 // 简单的内存实现，用于本地开发/测试和示例 wiring
 type InMemRepo struct {
 	mu   sync.RWMutex
-	data map[int64]appsdomain.Product
+	data map[int64]domain.Product
 	next int64
 }
 
-func NewInMemRepo() appsports.ProductRepository {
-	r := &InMemRepo{data: make(map[int64]appsdomain.Product), next: 1}
+func NewInMemRepo() ports.ProductRepo {
+	r := &InMemRepo{data: make(map[int64]domain.Product), next: 1}
 	// seed demo data
-	r.data[1] = appsdomain.Product{ID: 1, Name: "Blue Widget", Price: 1999}
-	r.data[2] = appsdomain.Product{ID: 2, Name: "Red Gizmo", Price: 2999}
+	r.data[1] = domain.Product{ID: 1, Name: "Blue Widget", Price: 1999}
+	r.data[2] = domain.Product{ID: 2, Name: "Red Gizmo", Price: 2999}
 	r.next = 3
 	return r
 }
 
-func (r *InMemRepo) GetByID(id int64) (*appsdomain.Product, error) {
+func (r *InMemRepo) GetByID(id int64) (*domain.Product, error) {
 	r.mu.RLock()
 	defer r.mu.RUnlock()
 	p, ok := r.data[id]
@@ -37,7 +37,7 @@ func (r *InMemRepo) GetByID(id int64) (*appsdomain.Product, error) {
 	return &pp, nil
 }
 
-func (r *InMemRepo) Search(q string, page, pageSize int) ([]appsdomain.Product, error) {
+func (r *InMemRepo) Search(q string, page, pageSize int) ([]domain.Product, error) {
 	if page < 1 {
 		page = 1
 	}
@@ -46,7 +46,7 @@ func (r *InMemRepo) Search(q string, page, pageSize int) ([]appsdomain.Product, 
 
 	r.mu.RLock()
 	defer r.mu.RUnlock()
-	var out []appsdomain.Product
+	var out []domain.Product
 	for _, p := range r.data {
 		if q == "" || strings.Contains(strings.ToLower(p.Name), q) {
 			out = append(out, p)
@@ -54,7 +54,7 @@ func (r *InMemRepo) Search(q string, page, pageSize int) ([]appsdomain.Product, 
 	}
 	// simple pagination
 	if start >= len(out) {
-		return []appsdomain.Product{}, nil
+		return []domain.Product{}, nil
 	}
 	end := start + pageSize
 	if end > len(out) {
@@ -63,7 +63,7 @@ func (r *InMemRepo) Search(q string, page, pageSize int) ([]appsdomain.Product, 
 	return out[start:end], nil
 }
 
-func (r *InMemRepo) Create(p *appsdomain.Product) (int64, error) {
+func (r *InMemRepo) Create(p *domain.Product) (int64, error) {
 	r.mu.Lock()
 	defer r.mu.Unlock()
 	id := r.next
