+++ apps/product-query-svc/adapters/postgres/product_repository.go
@@ -1,11 +1,11 @@
 package postgres
 
 import (
-	"github.com/fightingBald/GoTuto/apps/product-query-svc/ports"
 	internaldb "github.com/fightingBald/GoTuto/internal/adapters/db"
+	"github.com/fightingBald/GoTuto/internal/ports"
 	"github.com/jackc/pgx/v5/pgxpool"
 )
 
-func NewProductRepository(pool *pgxpool.Pool) ports.ProductRepository {
+func NewProductRepository(pool *pgxpool.Pool) ports.ProductRepo {
 	return internaldb.NewPGProductRepo(pool)
 }
