--- apps/product-query-svc/ports/product_repository_port.go
@@ -1,6 +1,0 @@
-package ports
-
-import ip "github.com/fightingBald/GoTuto/internal/ports"
-
-// Outbound port (alias to existing internal port to avoid duplication)
-type ProductRepository = ip.ProductRepo
